move stuff out of pages
