+++ packages/docs/src/components/TeamCards.tsx
@@ -4,8 +4,8 @@ import Jonny from "../about/images/Jonny Burger.png";
 import Mehmet from "../about/images/Mehmet Ademi.png";
 import Patric from "../about/images/Patric Salvisberg.png";
 import { TwitterLogo, LinkedInLogo, GitHubLogo, EmailLogo } from "./icons";
-import { BlueButton } from "../../../components/layout/Button";
-import { Spacer } from "../../../components/layout/Spacer";
+import { BlueButton } from "../../components/layout/Button";
+import { Spacer } from "../../components/layout/Spacer";
 
 const ButtonMailto = ({ mailto, label }) => {
   return <a href={mailto}>{label}</a>;
